+++ src/editors/ruleSetEditor.ts
@@ -1,17 +1,82 @@
+import * as vscode from 'vscode';
+import * as path from 'path';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
 import { JsonFormEditor } from "./jsonFormEditor";
+import { CARuleInfo } from "../langserver/caRuleInfo";
+import { ToolsGetCodeAnalyzersRulesRequest } from '../langserver/toolsGetCodeAnalyzersRulesRequest';
+import { basename } from 'path';
 
 export class RuleSetEditor extends JsonFormEditor {
+    protected _analyzers: string[];
 
     constructor(devToolsContext : DevToolsExtensionContext) {
         super(devToolsContext, "RuleSet");
 
         this.fieldsOrder = ["name", "description", "generalAction", "includedRuleSets", "rules"];
+        this._analyzers = ["${AppSourceCop}", "${CodeCop}", "${PerTenantExtensionCop}", "${UICop}"];
     }
 
     protected getViewType() : string {
         return 'azALDevTools.AppJsonEditor';
     }
+
+    protected onDocumentLoaded() {
+        super.onDocumentLoaded();
+        this.loadRules();
+    }
+
+    protected loadAnalyzers() {
+        let uri: vscode.Uri | undefined = undefined;
+        if (this.document)
+            uri = this.document.uri;
+        
+        let alConfig = vscode.workspace.getConfiguration('al', uri);
+        let codeAnalyzersSetting = alConfig.get<string[]|undefined>("codeAnalyzers");
+        if (codeAnalyzersSetting) {
+            for (let i=0; i<codeAnalyzersSetting.length; i++) {
+                if (!codeAnalyzersSetting[i].startsWith('${')) {
+                    this._analyzers.push(path.parse(codeAnalyzersSetting[i]).name);
+                }
+            }
+        }
+    }
+
+    protected async loadRules() {
+        await vscode.window.withProgress({
+            location: vscode.ProgressLocation.Notification,
+            title: "Loading code analyzers rules"
+        }, async (progress) => {
+            let rules: any[] = [];
+
+            this.loadAnalyzers();
+
+            for (let analyzerIdx=0; analyzerIdx<this._analyzers.length; analyzerIdx++) {
+                let request: ToolsGetCodeAnalyzersRulesRequest = 
+                    new ToolsGetCodeAnalyzersRulesRequest(this._analyzers[analyzerIdx]);
+                let response = 
+                    await this._devToolsContext.toolsLangServerClient.getCodeAnalyzersRules(request);
+                if ((response) && (response.rules)) {
+                    for (let ruleIdx = 0; ruleIdx < response.rules.length; ruleIdx++) {
+                        response.rules[ruleIdx].analyzer = this._analyzers[analyzerIdx];
+                        //this._rules.push(response.rules[ruleIdx]);
+                        if (response.rules[ruleIdx].id)
+                            rules.push({ 
+                                value: response.rules[ruleIdx].id,
+                                description: response.rules[ruleIdx].description
+                            });
+                    }
+                }
+            }
+
+            if (rules.length > 0) {
+                this.sendMessage({
+                    command : 'setAutocomplete',
+                    path: ['rules', 'id'],
+                    data : rules
+                });        
+            }
+        });
+    }    
 
     protected getFieldsDefinition(): any {
         return [